requirements.txt
asgiref             == 3.7.2
beautifulsoup4      == 4.12.2
certifi             == 2023.7.22
cffi                == 1.16.0
charset-normalizer  == 3.2.0
crispy-bootstrap5   == 2023.10
cryptography        == 41.0.7
defusedxml          == 0.7.1
Django              == 5.0
django-allauth      == 0.59.0
django-crispy-forms == 2.1
django-extensions   == 3.0.9
django-markdownx    == 4.0.7
gunicorn            == 20.0.4
idna                == 3.4
joblib              == 1.3.1
Markdown            == 3.5.1
numpy               == 1.25.2
oauthlib            == 3.2.2
pandas              == 2.0.3
Pillow              == 10.0.0
pip                 == 23.0.1
pycparser           == 2.21
PyJWT               == 2.8.0
python-dateutil     == 2.8.2
python3-openid      == 3.2.0
pytz                == 2023.3
requests            == 2.31.0
requests-oauthlib   == 1.3.1
scikit-learn        == 1.3.0
scipy               == 1.11.1
setuptools          == 65.5.0
six                 == 1.16.0
soupsieve           == 2.5
sqlparse            == 0.4.4
threadpoolctl       == 3.2.0
typing_extensions   == 4.7.1
tzdata              == 2023.3
urllib3             == 2.0.4